Rename ViewpoinToolBar.py to ViewpointTools.py
