+++ silx/gui/plot3d/tools/ViewpointTools.py
